meai_langchain/requirements.txt
langchain==0.0.314
langchain-openai==0.0.4
langchain-community==0.0.12
langchain-experimental==0.0.42
langchain-aws==0.0.3
pydantic==2.4.2
boto3==1.28.62
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
